Switch to Stack LTS release
This is currently broken with:

/usr/include/c++/5/bits/c++0x_warning.h:32:2: error: #error This file
requires compiler and library support for the ISO C++ 2011 standard.
This support must be enabled with the -std=c++11 or -std=gnu++11
compiler options.

I thought this was fixed by https://github.com/haskell/cabal/pull/4810
(which is in Cabal 2.2), but it might be broken because this:
https://github.com/haskell/cabal/pull/5315
(which is in is in Cabal 2.4, which isn't yet in LTS).